Project Update

+++ app/Http/Controllers/Admin/Settings/StaffController.php
@@ -2,9 +2,15 @@
 
 namespace App\Http\Controllers\Admin\Settings;
 
+use Carbon\Carbon;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Spatie\Permission\Models\Role;
 use App\Http\Controllers\Controller;
 use App\Models\Admin\Settings\Staff;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
 use Yajra\DataTables\Facades\DataTables;
 
 class StaffController extends Controller
@@ -31,7 +37,8 @@ class StaffController extends Controller
     public function create()
     {
         try {
-            return view('admin.settings.staff.create');
+            $roles = Role::orderBy('id', 'desc')->get();
+            return view('admin.settings.staff.create', compact('roles'));
         } catch (\Exception $exception) {
             return redirect()->back()->with('error', $exception->getMessage());
         }
@@ -45,18 +52,17 @@ class StaffController extends Controller
      */
     public function store(Request $request)
     {
-        // dd($request->all());
         $messages = array(
             'name.required' => 'Enter your name',
             'birth_date.required' => 'Enter birth date',
             'join_date.required' => 'Enter join date',
             'gender.required' => 'Choose a gender',
-            'image.required' => 'Upload a profile picture',
             'designation.required' => 'Enter your designation',
             'salary.required' => 'Enter your salary',
             'contact_no.required' => 'Enter your contact number',
             'email.required' => 'Enter your email',
             'address.required' => 'Enter your address',
+            'password.required' => 'Create password',
         );
 
         $this->validate($request, array(
@@ -65,17 +71,22 @@ class StaffController extends Controller
             'join_date' => 'required|',
             'gender' => 'required|',
             'designation' => 'required|',
-            'salary' => 'required|',
-            'contact_no' => 'required|',
-            'email' => 'required|',
-            'image.*' => 'max:2048'
+            'salary' => 'required|numeric',
+            'contact_no' => 'required|regex:/^([0-9\s\-\+\(\)]*)$/|min:11|unique:staff,contact_no,NULL,id,deleted_at,NULL',
+            'email' => 'required|string|unique:staff,email,NULL,id,deleted_at,NULL',
+            'image.*' => 'max:2048',
+            'password' => 'required|min:6',
+
         ), $messages);
+
+
+        DB::beginTransaction();
 
         try {
             $data = new Staff();
             $data->name = $request->name;
-            $data->birth_date = $request->birth_date;
-            $data->join_date = $request->join_date;
+            $data->birth_date = Carbon::parse($request->birth_date)->format('Y-m-d');
+            $data->join_date = Carbon::parse($request->join_date)->format('Y-m-d');
             $data->gender = $request->gender;
 
             if ($request->file('image')) {
@@ -88,16 +99,36 @@ class StaffController extends Controller
             $data->contact_no = $request->contact_no;
             $data->designation = $request->designation;
             $data->salary = $request->salary;
+            $data->password = Hash::make($request->password);
             $data->email = $request->email;
             $data->address = $request->address;
             $data->status = $request->status;
             $data->description = $request->description;
-            // dd($data);
+            $data->role = 3;
             $data->save();
+
+            $user = new User();
+            $user->staff_id = $data->id;
+            $user->name = $request->name;
+            $user->type = 3;
+            $user->status = 1;
+            $user->email  = $request->email;
+            $user->password  = Hash::make($request->password);
+            $user->save();
+            $user->assignRole($request->role);
+
+            DB::table('model_has_roles')->insert([
+                'role_id' => 3,
+                'model_type' => 'App\\Models\\User',
+                'model_id' => $user->id,
+            ]);
+
+            DB::commit();
 
             return redirect()->route('admin.staff.index')
             ->with('message', 'Staff created successfully');
         } catch (\Exception $exception) {
+            DB::rollback();
             return redirect()->back()->with('error', $exception->getMessage());
         }
     }
@@ -112,46 +143,66 @@ class StaffController extends Controller
     public function staff(Request $request)
     {
         try {
-            //--- Integrating This Collection Into Datatables
             if ($request->ajax()) {
-
                 $data = Staff::orderBy('id', 'desc')->get();
-
                 return Datatables::of($data)
                     ->addColumn('status', function ($data) {
+                        if (Auth::user()->can('staff_status')) {
                         $button = ' <div class="custom-control custom-switch">';
                         $button .= ' <input type="checkbox" class="custom-control-input changeStatus" id="customSwitch' . $data->id . '" getId="' . $data->id . '" name="status"';
 
                         if ($data->status == 1) {
-
                             $button .= "checked";
                         }
                         $button .= '><label for="customSwitch' . $data->id . '" class="custom-control-label" for="switch1"></label></div>';
                         return $button;
+                    }
+                    else{
+                        return "--";
+                    }
                     })
 
                     ->addColumn('gender', function (Staff $data) {
                         if ($data->gender == 1) {
-                            return '<p">
-                                        Male
-                                    </p>';
-                        } else {
-                            return '<p">
-                                   Female
-                               </p>';
+                            return 'Male';
+                        } elseif($data->gender == 2) {
+                            return 'Female';
+                        }else{
+                            return 'Other';
                         }
                     })
 
                     ->addColumn('action', function (Staff $data) {
-                        return ' <a href="' . route('admin.staff.show', $data->id) . ' " class="btn btn-sm btn-primary"><i class="fa fa-eye"></i></a>
 
-                       <a href="' . route('admin.staff.edit', $data->id) . ' " class="btn btn-sm btn-info"><i class="fa fa-edit"></i></a>
+                        if (Auth::user()->can('staff_show')) {
+                            $details =  '<a href="' . route('admin.staff-profile', $data->id) . ' " class="btn btn-sm btn-info" title="Profile"><i class="fa fa-user" aria-hidden="true"></i></a> ';
+                         } else {
+                                 $details = " ";
+                         }
 
-                    <button id="messageShow" class="btn btn-sm btn-danger btn-delete" data-remote=" ' . route('admin.staff.destroy', $data->id) . ' " title="Delete" ><i class="fa fa-trash-alt"></i></button>';
+                        if (Auth::user()->can('staff_show')) {
+                        $show = ' <a href="' . route('admin.staff.show', $data->id) . ' " class="btn btn-sm btn-primary" title="Show"><i class="fa fa-eye"></i></a> ';}
+                        else{
+                            $show = " ";
+                        }
+                        if (Auth::user()->can('staff_edit')) {
+                       $edit=' <a href="' . route('admin.staff.edit', $data->id) . ' " class="btn btn-sm btn-info" title="Edit"><i class="fa fa-edit"></i></a> ';}
+                       else{
+                        $edit = " ";
+                       }
+
+                       if (Auth::user()->can('staff_delete')) {
+                    $delete = ' <button id="messageShow" class="btn btn-sm btn-danger btn-delete" data-remote=" ' . route('admin.staff.destroy', $data->id) . ' " title="Delete" ><i class="fa fa-trash-alt"></i></button> ';}
+                    else{
+                        $delete = " ";
+                    }
+                    return $details.$show.$edit.$delete;
+
                     })
 
+                    ->addIndexColumn()
                     ->rawColumns(['status', 'action', 'gender'])
-                    ->toJson(); //--- Returning Json Data To Client Side
+                    ->toJson();
             }
         } catch (\Exception $exception) {
             return redirect()->back()->with('error', $exception->getMessage());
@@ -162,7 +213,8 @@ class StaffController extends Controller
     {
         try {
             $data = Staff::findOrFail($id);
-            return view('admin.settings.staff.show', compact('data'));
+            $roles = Role::orderBy('id', 'desc')->get();
+            return view('admin.settings.staff.show', compact('data', 'roles'));
         } catch (\Exception $exception) {
             return redirect()->back()->with('error', $exception->getMessage());
         }
@@ -178,7 +230,8 @@ class StaffController extends Controller
     {
         try {
             $data = Staff::findOrFail($id);
-            return view('admin.settings.staff.edit', compact('data'));
+            $roles = Role::orderBy('id', 'desc')->get();
+            return view('admin.settings.staff.edit', compact('data', 'roles'));
         } catch (\Exception $exception) {
             return redirect()->back()->with('error', $exception->getMessage());
         }
@@ -193,7 +246,6 @@ class StaffController extends Controller
      */
     public function update(Request $request, $id)
     {
-        // dd($request->all());
         $messages = array(
             'name.required' => 'Enter your name',
             'birth_date.required' => 'Enter birth date',
@@ -208,22 +260,24 @@ class StaffController extends Controller
         );
 
         $this->validate($request, array(
-            'name' => 'required|string|',
+            'name' => 'required|',
             'birth_date' => 'required|',
             'join_date' => 'required|',
             'gender' => 'required|',
             'designation' => 'required|',
             'salary' => 'required|',
-            'contact_no' => 'required|',
-            'email' => 'required|',
+            'contact_no' => 'required|regex:/^([0-9\s\-\+\(\)]*)$/|min:11|unique:staff,contact_no,' . $id . ',id,deleted_at,NULL',
+            'email' => 'required|unique:staff,email,' . $id . ',id,deleted_at,NULL',
             'image.*' => 'max:2048'
         ), $messages);
+
+        DB::beginTransaction();
 
         try {
             $data = Staff::findOrFail($id);
             $data->name = $request->name;
-            $data->birth_date = $request->birth_date;
-            $data->join_date = $request->join_date;
+            $data->birth_date = Carbon::parse($request->birth_date)->format('Y-m-d');
+            $data->join_date = Carbon::parse($request->join_date)->format('Y-m-d');
             $data->gender = $request->gender;
 
             // Store Image
@@ -239,15 +293,25 @@ class StaffController extends Controller
             $data->designation = $request->designation;
             $data->salary = $request->salary;
             $data->email = $request->email;
+            $data->password = Hash::make($request->password);
             $data->address = $request->address;
             $data->status = $request->status;
             $data->description = $request->description;
-            // dd($data);
             $data->update();
+
+            $user = User::where('staff_id', $id)->first();
+            $user->staff_id = $data->id;
+            $user->name = $request->name;
+            $user->email  = $request->email;
+            $user->password  = Hash::make($request->password);
+            $user->update();
+            $user->syncRoles($request->role);
+            DB::commit();
 
             return redirect()->route('admin.staff.index')
             ->with('message', 'Staff updated successfully');
         } catch (\Exception $exception) {
+            DB::rollback();
             return redirect()->back()->with('error', $exception->getMessage());
         }
     }
@@ -258,22 +322,27 @@ class StaffController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+
     public function destroy($id)
     {
         try {
-            $data = Staff::findOrFail($id);
-            $data->delete();
-            return back()->with('message', 'Data deleted successfully');
+            $data =  Staff::findOrFail($id);
+            if ($data) {
+                $user = User::where('staff_id', $id)->first();
+                DB::table('model_has_roles')->where('role_id', 3)->where('model_id', $user->id)->delete();
+                $user->delete();
+                $data->delete();
+                return redirect()->route('admin.staff.index')
+                ->with('message', 'Staff deleted successfully');
+            }
         } catch (\Exception $exception) {
             return redirect()->back()->with('error', $exception->getMessage());
         }
     }
 
-    // change status function start here
     public function StatusChange(Request $request)
     {
         $id = $request->id;
-
         $status_check   = Staff::findOrFail($id);
         $status         = $status_check->status;
 
@@ -288,10 +357,69 @@ class StaffController extends Controller
         Staff::where('id', $id)->update($data);
         if ($status_update == 1) {
             return "success";
-            exit();
         } else {
             return "failed";
         }
     }
-    // end change function
+
+    public function StaffProfile($id)
+    {
+        try {
+            $data = Staff::findOrFail($id);
+            $roles = Role::orderBy('id', 'desc')->get();
+            return view('admin.settings.staff.staff_profile', compact('data', 'roles'));
+        } catch (\Exception $exception) {
+            return redirect()->back()->with('error', $exception->getMessage());
+        }
+    }
+
+    public function StaffProfileUpdate(Request $request, $id){
+        if($request->password != $request->confirm_password){
+            return redirect()->back()->with('error', "Passwords do not match");
+        }
+
+        if($request->password){
+            $messages = array(
+                'password' => 'Password must be be minimum 6 digit',
+            );
+
+            $this->validate($request, array(
+                'password' => 'min:6',
+            ), $messages);
+        }
+
+        DB::beginTransaction();
+
+        try {
+            $data = Staff::findOrFail($id);
+            $data->name = $request->name;
+            $data->contact_no = $request->contact_no;
+            $data->email = $request->email;
+            $data->address = $request->address;
+            $data->password = Hash::make($request->password);
+            // Store Image
+            if (request()->has('image')) {
+                @unlink(public_path('img/') . $data->image);
+                $file = $request->file('image');
+                $filename = time() . $file->getClientOriginalName();
+                $file->move(public_path('/img/'), $filename);
+                $data->image = $filename;
+            }
+            $data->update();
+
+            $user = User::where('staff_id', $id)->first();
+            $user->staff_id = $data->id;
+            $user->name = $request->name;
+            $user->email  = $request->email;
+            $user->password = Hash::make($request->password);
+            $user->update();
+
+            DB::commit();
+
+            return redirect()->back()->with('message', 'Data updated successfully');
+        } catch (\Exception $exception) {
+            DB::rollback();
+            return redirect()->back()->with('error', $exception->getMessage());
+        }
+    }
 }
